Add single test ability to checktsh.py
checktsh.py now can take in a command line argument, the trace to run.
If the trace is provided, run it, exiting with status 0 if the trace
passed, and exiting with status 1 if the trace failed.

If a trace is not passed on the command line, run all traces and print
their results and final score.